outsourced ModdleCopy
