+++ app/lib/rolemodeler/features/copy-paste/index.js
@@ -1,7 +1,7 @@
 import CopyPasteModule from 'diagram-js/lib/features/copy-paste';
 
 import ODCopyPaste from './ODCopyPaste';
-import ModdleCopy from './ModdleCopy';
+import ModdleCopy from '../../../frequentlyUsedFiles/features/copy-paste/ModdleCopy.js';
 
 export default {
   __depends__: [
